Re-add gir-repository-dev to build-dependencies
The Soup-2.4.gir file has still not been moved from
gir-repository-dev to libsoup2.4-dev in unstable (only experimental
so far).

Since the gir-repository-dev package was updated to gir1.2 in
version 0.6.5-9, lets add a dependency on that for now....

We can drop the g-r-dev b-d when the gir has been moved.
At the same time also add a version (>= 2.32.2-1) to the
libsoup2.4-dev build dependency.