Replace gupnp-vala before 0.10.5 to avoid conflicts
Since gupnp now builds it's own vapi and deps from gir,
we need to declare a replace against previous versions of
gupnp-vala which used to ship these files.
From version 0.10.5 of gupnp-vala, gupnp (and gupnp-av)
vala api will not be shipped if built against new
enough gupnp (and gupnp-av) versions. A versioned
build-dependency dependency should be added in
the gupnp-vala package to make sure they are not
built/shipped anymore.