Drop explicit build-dependencies on gir1.2-*
The new gobject introspection mini policy says that the -dev
packages should depend on their gir's, to simplify build-dependencies.
We enabled gir before that, but lets now use this simplification...

Bump libgirepository1.0-dev to (>= 0.10.7-1~) for the new policy
and getting the gir1.2-freedesktop and gir1.2-glib-2.0 dependencies.

Drop gir1.2-gssdp-1.0 and gir1.2-soup-2.4 since these are already pulled
in by their respective -dev packages which we already build-depend on.